Update aUtils.vs
added a check to see if diob that is being transitioned is an object. If so it will need to store the value differently. (This is for usability in all cases, and not just using it with pure diobs, but your own object tracker variables you can then use to apply the color to something like a shader.)