`fix & perf`: Fixed a bug, and stopped the recreation of an object
This library has been updated to be better support the `aParticleGenerator` library. When a particle's generator is paused, the transitioning of it's color will now be paused as well.
This library no longer recreates the `color` object each time a new color is given. It reuses the object and just manually updates the `color` object with itself so it can render the change. (This is due to the engine not listening to the `tint` variable inside of `color`)