fix
Fixed a dormant bug that caused the aMetaballsLighting library to twitch with its colors. This bug was that if an object was passed to aUtils to have its tint variable changed over time to another color, it would not assign the data to color.tint, but rather change the color object itself causing error. This fix changes it.