handle event where a diob is deleted
When using the transitionColor function on a particle that comes from the particle generator plugin it may be deleted while it is being transitioned. The aUtils library previously did not handle this properly and caused errors.
This change implements clearing the interval started via transitionColor and removing it from tracking.